Proof of concept of package analysis using GHunter
Add two examples of using GHunter to analyze packages for gadgets. This
is achieved quite simply by telling the `analyze.sh` script about these
new targets and their test directory. The rest is handled inherently by
the tooling.

It should be noted that this is not generally the case. These libraries
follow conventions also used by the Deno project itself, meaning
assumptions made in GHunter tooling hold. Other libraries may be a bit
harder to adopt, e.g. because they don't use the `_test.ts` suffix for
tests.

The two packages covered are Deno's Fresh framework and the
`denodrivers/postgres` project. For the former we find no gadgets
while for the latter we do find a few, proving that this actually
works.